test: extract shared FEBase page object and add fast smoke script
Phase 3 (simpler/​DRY test setup), examples test infra only — no component/src changes:

- new FEBase holds the parts that are genuinely identical across FEV2 / FEV4 /
  FEV2ND (unicode-space regex, Base helper, nav/lookup constants, getTableItems
  /getTableObject/getTimeValue). FEV2, FEV4 and FEV2ND now `extends FEBase` and
  keep only their template/version-specific selectors and the field/date/routing
  getters that depend on them. Public surface of each class is unchanged
  (verified: identical instance keys + string-prop values + methods; FEV4 only
  gains the inherited getTimeValue util).
- drop the dead `require("./FE")` from the three page objects.
- add `npm run test:smoke` — single-spec V2 deep-export/update/draft run on the
  default UI5 version for fast local feedback without the full matrix; documented
  in CLAUDE.md.
