+++ CLAUDE.md
@@ -126,7 +126,8 @@ npm install --legacy-peer-deps   # install (monorepo, npm workspaces)
 npm run test:unit                # Jest unit tests (no server)
 npm run start:server             # CAP server :4004 (cds watch, OData v2+v4)
 npm run start:v2fe               # V2 FE app :8081   (start:v4fe → :8080)
-npm run test:v2fe                # wdi5 V2 FE (server + app must be running)
+npm run test:v2fe                # wdi5 V2 FE, full scenario (server + app must be running)
+npm run test:smoke               # wdi5 V2 deep-export/update/draft, single spec — fast local check
 npm run lint                     # eslint
 npm run prettier                 # format
 # DO NOT run `npm run build` for local dev/test — it populates dist/ (see top).
